MSD203-45 Test cases for UserServie and GroupService #comment compatible with new code

+++ Project_src/Prattle/src/main/java/edu/northeastern/ccs/im/service/GroupServicePrattle.java
@@ -1,21 +1,19 @@
 package edu.northeastern.ccs.im.service;
 
-import java.util.List;
-
-import com.mongodb.client.model.Filters;
-import com.mongodb.client.result.DeleteResult;
-import com.mongodb.client.FindIterable;
-import com.mongodb.client.MongoCollection;
-import com.mongodb.client.MongoDatabase;
-import com.mongodb.client.model.Updates;
-import org.bson.Document;
-
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.gson.Gson;
-
+import com.mongodb.client.FindIterable;
+import com.mongodb.client.MongoCollection;
+import com.mongodb.client.MongoDatabase;
+import com.mongodb.client.model.Filters;
+import com.mongodb.client.model.Updates;
+import com.mongodb.client.result.DeleteResult;
 import edu.northeastern.ccs.im.MongoDB.Model.Group;
 import edu.northeastern.ccs.im.MongoDB.Model.User;
+import org.bson.Document;
+
+import java.util.List;
 
 /**
  *
@@ -78,8 +76,12 @@ public class GroupServicePrattle {
 		List<String> listOfUsers = findGroupByName(groupname).getListOfUsers();
 		DeleteResult dr =gcol.deleteOne(Filters.eq("name", groupname.toLowerCase()));
 		boolean removedGroup=false;
-		for(String username: listOfUsers){
-			removedGroup=userService.removeGroupFromUser(username,groupname.toLowerCase());
+		if(listOfUsers.size()==0){
+			removedGroup=true;
+		}else {
+			for (String username : listOfUsers) {
+				removedGroup = userService.removeGroupFromUser(username, groupname.toLowerCase());
+			}
 		}
 
 		return (dr.wasAcknowledged() && removedGroup);
